refactor: Public pages

+++ src/Admin/Page/Application/Gateway/Delete/Gateway.js
@@ -1,12 +1,17 @@
 const Gateway = require('../../../../../Shared/Application/Gateway/Gateway');
 
-module.exports = class CreateGateway extends Gateway {
+module.exports = class DeleteGateway extends Gateway {
     constructor(repository) {
         super('admin.pages.delete');
         this.repository = repository;
     }
 
     async process({ id }) {
+        const page = await this.repository.get(id);
+        if (page.isFactory) {
+            throw new Error(`Page ${page.slug} cannot be deleted because is factory resource.`);
+        }
+
         return await this.repository.remove(id);
     }
 }
